basic arithmetic

+++ app/src/main/java/com/example/kotlinbasic/kotlinprogramms/Operators.kt
@@ -6,5 +6,5 @@ fun main() {
     println(a+b)
     println(a-b)
     println(a.toFloat()/b)
-    println(a.toFloat()*b)
+    println(a*b)
 }
